rename withSettingsts => settings.ts, use getSettings instead of withSettings
